fix(systemd-ask-password): do not half-install systemd-ask-password-wall
Do not install the path unit when the service unit is not installed
for systemd-ask-password-wall.

Fixes the following warning on the CI:

[FAILED] Failed to start Forward Password Requests to Wall Directory Watch.
See 'systemctl status systemd-ask-password-wall.path' for details.
...
systemd[1]: systemd-ask-password-wall.path: Refusing to start, unit systemd-ask-password-wall.service to trigger not loaded.
systemd[1]: Failed to start Forward Password Requests to Wall Directory Watch.